requirements.txt
websockets
misskey.py
speedtest-cli
python-dotenv
